+++ packages/core/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The following is a curated list of changes in the Enact core module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `core/util.memoize` method to memoize functions
+
+### Removed
+
+- `core/util.childrenEquals` which was no longer supported by React 16
+
 ## [2.0.0-alpha.4] - 2018-02-13
 
 ### Changed
@@ -20,7 +30,19 @@ No significant changes.
 
 No significant changes.
 
-## [1.13.3] - 2017-01-16
+## [1.15.0] - 2018-02-28
+
+### Deprecated
+
+- `core/util/childrenEquals`, to be removed in 2.0.0
+
+## [1.14.0] - 2018-02-23
+
+### Deprecated
+
+- `core/factory`, to be removed in 2.0.0
+
+## [1.13.3] - 2018-01-16
 
 ### Fixed